feat: add real-time game session with challenge routing and scoring
- Add option shuffling with independent per-player permutations per round
- Add GameSession goroutine: 10-round loop with 15s timeout, RTT-compensated scoring
- Add challenge routing in hub via GameStarter interface (avoids import cycle)
- Add game.Starter that creates matches, loads questions, and spawns sessions
- Add game metrics: active/total games, rounds, fair time, latency, score distribution
- Fix UpdateBothRatings query with explicit int cast for Postgres CASE
- Relax ChallengeRespMsg fields to strings (only MatchID and Accepted needed)
- Add shuffle unit tests and full game integration test (10 rounds end-to-end)
- All key invariants verified: consistent scores, independent shuffles, atomic ratings

+++ internal/db/queries.sql.go
@@ -731,8 +731,8 @@ func (q *Queries) SearchUsersByUsername(ctx context.Context, dollar_1 pgtype.Tex
 
 const updateBothRatings = `-- name: UpdateBothRatings :exec
 UPDATE users SET rating = CASE
-    WHEN id = $1 THEN $2
-    WHEN id = $3 THEN $4
+    WHEN id = $1 THEN $2::int
+    WHEN id = $3 THEN $4::int
 END
 WHERE id IN ($1, $3)
 `
